+++ src/GameSvr/Event/Events/HolyCurtainEvent.cs
@@ -4,10 +4,9 @@ namespace GameSvr.Event.Events
 {
     public class HolyCurtainEvent : EventInfo
     {
-        public HolyCurtainEvent(Envirnoment Envir, int nX, int nY, int nType, int nTime) : base(Envir, nX, nY, nType, nTime, true)
+        public HolyCurtainEvent(Envirnoment Envir, short nX, short nY, byte nType, int nTime) : base(Envir, nX, nY, nType, nTime, true)
         {
 
         }
     }
 }
-
